# public/swagger.json
{
	"components": {
...
		"schemas": {
			"ResponseData": {
				"properties": {
					"status": {
						"type": "number",
						"enum": [0, 1],
						"description": "1 = succeed | 0 = failed"
					},
					"data": {},
					"messages": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Error/warning messages"
					}
				},
				"required": ["status", "data", "messages"],
				"type": "object",
				"additionalProperties": false
			},
...
			"ObjectID": {
				"description": "Create a new ObjectID instance.",
				"properties": {
					"generationTime": {
						"type": "number",
						"format": "double",
						"description": "The generation time of this ObjectId instance."
					}
				},
				"required": ["generationTime"],
				"type": "object",
				"additionalProperties": false
			},
...
			"ProviderInfo": {
				"properties": {
					"name": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					},
					"access_token": {
						"type": "string"
					}
				},
				"required": ["name"],
				"type": "object",
				"additionalProperties": false
			},
			"AccessTokenInfo": {
				"properties": {
					"access_token": {
						"type": "string"
					},
					"expiredTimestamp": {
						"type": "number",
						"format": "double"
					},
					"expiredDate": {
						"type": "string",
						"format": "date-time"
					},
					"expiredDateGTM7": {
						"type": "string"
					}
				},
				"required": ["access_token", "expiredTimestamp", "expiredDate", "expiredDateGTM7"],
				"type": "object",
				"additionalProperties": false
			},
			"IRouteScope": {
				"type": "string",
				"enum": ["all", "workspace", "team", "app", "own"]
			},
			"IRoutePermission": {
				"type": "string",
				"enum": ["full", "create", "read", "update", "delete"]
			},
			"RoleRoute": {
				"properties": {
					"route": {
						"type": "string"
					},
					"scope": {
						"$ref": "#/components/schemas/IRouteScope"
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/IRoutePermission"
						},
						"type": "array"
					}
				},
				"required": ["route", "scope", "permissions"],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string",
						"description": "User name"
					},
					"username": {
						"type": "string",
						"description": "Unique username of a user\nThis equavilent with \"slug\""
					},
					"type": {
						"type": "string",
						"enum": ["user", "service_account"],
						"description": "Service Account is User!"
					},
					"email": {
						"type": "string",
						"description": "User email address"
					},
					"verified": {
						"type": "boolean",
						"description": "Is this user's email or phone verified?"
					},
					"image": {
						"type": "string",
						"description": "User profile picture URL"
					},
					"providers": {
						"items": {
							"$ref": "#/components/schemas/ProviderInfo"
						},
						"type": "array",
						"description": "List of Cloud Providers which this user can access to"
					},
					"password": {
						"type": "string",
						"description": "User password (hashed)"
					},
					"token": {
						"$ref": "#/components/schemas/AccessTokenInfo",
						"description": "User token"
					},
					"roles": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/ObjectID"
								},
								{
									"$ref": "#/components/schemas/Role"
								},
								{
									"type": "string"
								}
							]
						},
						"type": "array",
						"description": "User's roles\nTODO: Thêm \"workspace\" cho \"roles\" vì mỗi workspace sẽ có roles khác nhau, trong khi mỗi user có thể nằm trong nhiều workspace"
					},
					"teams": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/ObjectID"
								},
								{
									"$ref": "#/components/schemas/Team"
								},
								{
									"type": "string"
								}
							]
						},
						"type": "array",
						"description": "User's team IDs which this user is a member"
					},
					"workspaces": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/ObjectID"
								},
								{
									"$ref": "#/components/schemas/Workspace"
								},
								{
									"type": "string"
								}
							]
						},
						"type": "array",
						"description": "List of workspace IDs which this user is a member"
					},
					"activeWorkspace": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Workspace"
							},
							{
								"type": "string"
							}
						],
						"description": "Active workspace of a user"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"ObjectId": {
				"type": "string",
				"description": "A class representation of the BSON ObjectId type."
			},
			"ResourceQuotaSize": {
				"type": "string",
				"enum": ["none", "1x", "2x", "3x", "4x", "5x", "6x", "7x", "8x", "9x", "10x"]
			},
			"SslType": {
				"type": "string",
				"enum": ["none", "letsencrypt", "custom"],
				"default": "\"letsencrypt\""
			},
			"DeployEnvironment": {
				"description": "A deployment's environment of the application.",
				"properties": {
					"imageURL": {
						"type": "string",
						"description": "Image URI of this app on the Container Registry.\n- Combined from: `<registry-image-base-url>/<project-slug>/<app-name-slug>`\n- If you build from the source code, don't specify `tag` at the end! (eg. `latest`, `beta`,...)"
					},
					"namespace": {
						"type": "string",
						"description": "Destination namespace name"
					},
					"registry": {
						"type": "string",
						"description": "Container registry slug"
					},
					"provider": {
						"type": "string",
						"description": "Cloud provider's short name"
					},
					"cluster": {
						"type": "string",
						"description": "Cluster's short name"
					},
					"project": {
						"type": "string",
						"description": "[Google Cloud] PROJECT_ID",
						"deprecated": true
					},
					"region": {
						"type": "string",
						"description": "[Google Cloud] Region",
						"deprecated": true
					},
					"zone": {
						"type": "string",
						"description": "[Google Cloud] Zone",
						"deprecated": true
					},
					"size": {
						"$ref": "#/components/schemas/ResourceQuotaSize",
						"description": "Container quota resources"
					},
					"shouldInherit": {
						"type": "boolean",
						"description": "Set to `false` if you DON'T want to inherit the Ingress YAML config from the previous deployment",
						"default": "true"
					},
					"redirect": {
						"type": "boolean",
						"description": "Set to `false` if you don't want to redirect all the secondary domains to the primary domain.",
						"default": "true"
					},
					"replicas": {
						"type": "number",
						"format": "double",
						"description": "Container's scaling replicas",
						"default": "1"
					},
					"port": {
						"type": "number",
						"format": "double",
						"description": "Container's port"
					},
					"basePath": {
						"type": "string",
						"description": "Application base path in the endpoint URL",
						"default": "\"/\""
					},
					"domains": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "List of application's domains"
					},
					"cdn": {
						"type": "boolean",
						"description": "Flag to enable CDN for this application",
						"default": "false;"
					},
					"ssl": {
						"$ref": "#/components/schemas/SslType",
						"description": "SSL Certificate Issuer"
					},
					"tlsSecret": {
						"type": "string",
						"description": "Secret name to hold the key of SSL, will be automatically generated.\nOnly need to specify when using \"custom\" SSL (which is the SSL from third-party issuer)"
					},
					"ingress": {
						"type": "string",
						"description": "Kubernetes Ingress Class"
					},
					"cliVersion": {
						"type": "string",
						"description": "The CLI version"
					},
					"namespaceYaml": {
						"type": "string",
						"description": "Content of namespace YAML file"
					},
					"deploymentYaml": {
						"type": "string",
						"description": "Content of deployment YAML file"
					},
					"prereleaseDeploymentYaml": {
						"type": "string",
						"description": "Content of prerelease deployment YAML file"
					},
					"prereleaseUrl": {
						"type": "string",
						"description": "Prerelease endpoint URL"
					},
					"envVars": {
						"items": {
							"$ref": "#/components/schemas/KubeEnvironmentVariable"
						},
						"type": "array",
						"description": "Collection array of environment variables"
					},
					"createdBy": {
						"type": "string",
						"description": "User name of the first person who deploy on this environment."
					},
					"lastUpdatedBy": {
						"type": "string",
						"description": "User name of the last person who deploy or update this environment."
					},
					"creator": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectId"
							},
							{
								"$ref": "#/components/schemas/User"
							}
						],
						"description": "ID of the creator"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "Update time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Project": {
				"description": "Projects",
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"apiKey": {
						"type": "string"
					},
					"clientId": {
						"type": "string"
					},
					"clientSecret": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"lastUpdatedBy": {
						"type": "string"
					},
					"latestBuild": {
						"type": "string"
					},
					"appSlugs": {
						"type": "string"
					},
					"apps": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/ObjectID"
								},
								"type": "array"
							},
							{
								"items": {
									"$ref": "#/components/schemas/App"
								},
								"type": "array"
							},
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						],
						"description": "List of App IDs"
					},
					"owner": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/User"
							},
							{
								"type": "string"
							}
						],
						"description": "User ID of the owner"
					},
					"workspace": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Workspace"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the workspace"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Role": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"routes": {
						"items": {
							"$ref": "#/components/schemas/RoleRoute"
						},
						"type": "array"
					},
					"owner": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/User"
							},
							{
								"type": "string"
							}
						],
						"description": "User ID of the owner"
					},
					"project": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Project"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the project"
					},
					"workspace": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Workspace"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the workspace"
					}
				},
				"required": ["name", "routes"],
				"type": "object",
				"additionalProperties": false
			},
			"WorkspaceApiAccessToken": {
				"properties": {
					"name": {
						"type": "string"
					},
					"slug": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/Role"
						},
						"type": "array"
					}
				},
				"required": ["name", "token"],
				"type": "object",
				"additionalProperties": false
			},
			"Workspace": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Workspace slug: auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string",
						"description": "Workspace name"
					},
					"public": {
						"type": "boolean",
						"description": "Is this a `public` workspace that anyone can join, otherwise it's a `private` workspace that needs approval on joining request."
					},
					"image": {
						"type": "string",
						"description": "Workspace profile picture"
					},
					"domain": {
						"type": "string",
						"description": "Workspace domain name"
					},
					"apiAccessTokens": {
						"items": {
							"$ref": "#/components/schemas/WorkspaceApiAccessToken"
						},
						"type": "array",
						"description": "List of this workspace's API Access Token"
					},
					"owner": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/User"
							},
							{
								"type": "string"
							}
						],
						"description": "User ID of the owner"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"App": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"lastUpdatedBy": {
						"type": "string"
					},
					"git": {
						"properties": {
							"repoSSH": {
								"type": "string"
							},
							"repoURL": {
								"type": "string"
							},
							"provider": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"framework": {
						"properties": {
							"repoSSH": {
								"type": "string"
							},
							"repoURL": {
								"type": "string"
							},
							"slug": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"environment": {
						"properties": {},
						"additionalProperties": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/DeployEnvironment"
								},
								{
									"type": "string"
								}
							]
						},
						"type": "object",
						"deprecated": true
					},
					"deployEnvironment": {
						"properties": {},
						"additionalProperties": {
							"$ref": "#/components/schemas/DeployEnvironment"
						},
						"type": "object"
					},
					"latestBuild": {
						"type": "string"
					},
					"projectSlug": {
						"type": "string"
					},
					"owner": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/User"
							},
							{
								"type": "string"
							}
						],
						"description": "Owner ID of the app"
					},
					"project": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Project"
							},
							{
								"type": "string"
							}
						],
						"description": "Project ID of the app"
					},
					"workspace": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Workspace"
							},
							{
								"type": "string"
							}
						],
						"description": "Workspace ID of the app"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Team": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"type": "string"
							}
						]
					},
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"metadata": {
						"description": "Auto-generated by entity's columns\nEntity's metadata without unicode characters"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"owner": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/User"
							},
							{
								"type": "string"
							}
						],
						"description": "User ID of the owner"
					},
					"project": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Project"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the project"
					},
					"workspace": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ObjectID"
							},
							{
								"$ref": "#/components/schemas/Workspace"
							},
							{
								"type": "string"
							}
						],
						"description": "ID of the workspace"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"WorkspaceInputData": {
				"properties": {
					"slug": {
						"type": "string",
						"description": "Auto-generated by \"name\" column"
					},
					"active": {
						"type": "boolean",
						"description": "Auto-generated by \"name\" column"
					},
					"name": {
						"type": "string",
						"description": "Name of the workspace."
					},
					"owner": {
						"type": "string",
						"description": "User ID of the owner"
					},
					"public": {
						"type": "boolean",
						"description": "Set privacy mode for this workspace",
						"default": "true"
					}
				},
				"required": ["name", "owner"],
				"type": "object",
				"additionalProperties": false
			},
			"AddUserBody": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"workspaceId": {
						"type": "string"
					},
					"roleId": {
						"type": "string"
					}
				},
				"required": ["userId", "workspaceId"],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "API_ACCESS_TOKEN",
				"in": "header"
			},
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "@topgroup/diginext",
		"version": "2.6.5",
		"description": "A BUILD SERVER & CLI to deploy apps to any Kubernetes clusters.",
		"license": {
			"name": "GPL"
		},
		"contact": {
			"name": "TOP GROUP (a.k.a Digitop)",
			"email": "dev@wearetopgroup.com",
			"url": "https://github.com/digitopvn/diginext"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/workspace": {
			"get": {
				"operationId": "Read",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"anyOf": [
														{
															"$ref": "#/components/schemas/Workspace"
														},
														{
															"items": {
																"$ref": "#/components/schemas/Workspace"
															},
															"type": "array"
														}
													]
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Workspace"],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "populate",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "owner,workspace"
					},
					{
						"in": "query",
						"name": "select",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "-updatedAt,-createdAt"
					},
					{
						"description": "Disable the default `{$set: body}` of \"update\" query & update `{body}` directly to the items",
						"in": "query",
						"name": "raw",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "size",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Find one item by `{ObjectID}`",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "_id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Mark this request as search (return the similar results based on the filter query params)",
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"default": "true",
							"type": "boolean"
						}
					},
					{
						"description": "If `true`, return the excel binary file to download.",
						"in": "query",
						"name": "download",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseData"
								}
							}
						}
					}
				},
				"tags": ["Workspace"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/WorkspaceInputData"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"items": {
														"$ref": "#/components/schemas/Workspace"
													},
													"type": "array"
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Workspace"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "populate",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "owner,workspace"
					},
					{
						"in": "query",
						"name": "select",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "-updatedAt,-createdAt"
					},
					{
						"description": "Disable the default `{$set: body}` of \"update\" query & update `{body}` directly to the items",
						"in": "query",
						"name": "raw",
						"required": false,
						"schema": {
							"default": "false",
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/WorkspaceInputData"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"properties": {
														"n": {
															"type": "number",
															"format": "double"
														},
														"ok": {
															"type": "number",
															"format": "double"
														}
													},
													"type": "object"
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Workspace"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Delete one item by `{ObjectID}`",
						"in": "query",
						"name": "id",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Delete one item by `{slug}`",
						"in": "query",
						"name": "slug",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/workspace/add-user": {
			"patch": {
				"operationId": "AddUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ResponseData"
										},
										{
											"properties": {
												"data": {
													"items": {
														"$ref": "#/components/schemas/User"
													},
													"type": "array"
												}
											},
											"required": ["data"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": ["Workspace"],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddUserBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}